Eriytä adaptToDB:n commonin funktiot omiksi tiedostoikseen. Tee uusiksi adaptTiedostotToSave.
