+++ backend/src/projekti/adapter/adaptToDB/common/forEverySaameDoAsync.ts
@@ -1,0 +1,7 @@
+import { SaameKieli } from "../../../../database/model";
+
+export async function forEverySaameDoAsync(func: (kieli: SaameKieli) => Promise<void>): Promise<void> {
+  for (const saame in SaameKieli) {
+    await func(saame as SaameKieli);
+  }
+}
