+++ backend/src/projekti/adapter/adaptToDB/common/getId.ts
@@ -1,0 +1,14 @@
+export function getId(
+  vaihe:
+    | {
+        id: number | undefined;
+      }
+    | undefined
+    | null
+): number {
+  let id = vaihe?.id;
+  if (!id) {
+    id = 1;
+  }
+  return id;
+}
