+++ backend/src/projekti/adapter/adaptToDB/common/index.ts
@@ -1,0 +1,12 @@
+export * from "./adaptTiedostotToSave";
+export * from "./removeTypeName";
+export * from "./adaptLokalisoituTekstiEiPakollinen";
+export * from "./adaptIlmoituksenVastaanottajatToSave";
+export * from "./adaptYhteystiedotToSave";
+export * from "./adaptAineistotToSave";
+export * from "./adaptHankkeenKuvausToSave";
+export * from "./adaptLokalisoituTekstiToSave";
+export * from "./adaptLokalisoituLinkkiToSave";
+export * from "./getId";
+export * from "./forEverySaameDoAsync";
+export * from "./adaptLadattuTiedostoToSave";
